+++ NEWS.md
@@ -1,4 +1,6 @@
 # IDEATools 1.1
 
 * Added a `NEWS.md` file to track changes to the package.
+* Added a new function for PDF reporting : `MakeReport()`
+* Added a new function for Excel reporting : `MakeExcel()` which adds a new dependency to `{openxlsx}`
 * Added a new wrapper function, `DiagIDEA()` that allows users to complete a full IDEA diagnostic with only one line of code.